+++ CHANGELOG.md
@@ -609,3 +609,5 @@
 * Define `o_rwr_yio` operator
 * Define `pass` expression
 * Define `Statefully'` pattern
+* Define `In` pattern
+* Define `Out` pattern